// backend/src/types/types.ts
export enum Role {
  LAWYER = "LAWYER",
  CLIENT = "CLIENT",
}

export interface UserType {
  id: string;
  role: Role;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phone_number: string;
  address: string;
  addhar_number: string;
  profile: string;
  createdAt?: Date;
  updatedAt?: Date;
  Lawyer?: {
    profile_description?: string;
    yoe?: number;
    specialization?: string;
    license_number?: string;
    bar_association_membership?: string;
    consultationFee?: number;
    website?: string;
  };
}

export interface LawyerDetailsType {
  id: string;
  yoe: number;
  Ratings: Array<{
    id: string;
    client_id: string;
    rating: number;
    description: string;
  }>;
  specialization: string;
  consultationFee: number;
  createdAt: Date;
  updatedAt: Date;
  services: Array<{
    id: string;
    category: string;
    description: string;
    title: string;
    price: number;
  }>;
  user: {
    firstName: string;
    lastName: string;
    email: string;
    address: string;
  };
}

export interface AuthResponse {
  data?: any;
  error?: {
    message: string;
  };
}

export interface ServiceType {
  id?: string; // Optional, as it will be generated by the database
  title: string;
  description: string;
  category: string;
  price: number;
  createdAt?: string; // Optional, as it will be generated by the database
  updatedAt?: string; // Optional, as it will be updated automatically
  userId?: string;
}

export interface Rating {
  id?: string;
  lawyerId: string;
  clientId: string;
  description: string;
  rating: number;
}